"The computer should have support for large numbers"

+++ intcode/computer.c
@@ -9,31 +9,31 @@
 	#define trace(...)
 #endif
 
-void print_reel(size_t n, int reel[n]) {
+void print_reel(size_t n, long reel[n]) {
 	printf("[");
-	for (int *i = reel; i < reel + n; i++) {
-		printf("%d, ", *i);
+	for (long *i = reel; i < reel + n; i++) {
+		printf("%ld, ", *i);
 	}
 	printf("]\n");
 }
 
-int getval(int *r, int v, int **mode) {
+long getval(long *r, long v, int **mode) {
 	switch (*((*mode)++)) {
 		case 1:
-			trace(" %d ", v);
+			trace(" %ld ", v);
 			return v;
 		case 0:
-			trace(" *%d(=%d) ", v, r[v]);
+			trace(" *%ld(=%ld) ", v, r[v]);
 			return r[v];
 	}
 }
 
 #define lenof(l) sizeof(l) / sizeof(l[0])
-void computer(int *r, FILE *input) {
-	int *p = r;
-	int op;
+void computer(long *r, FILE *input) {
+	long *p = r;
+	long op;
 	trace("pc: %ld\n", (p) - r);
-	trace("op: %d\n", *p);
+	trace("op: %ld\n", *p);
 	while ((op = *(p++))) {
 		int pmode_t[8] = {0};
 		int *pmode = pmode_t;
@@ -43,29 +43,29 @@ void computer(int *r, FILE *input) {
 		switch (op % 100) {
 			case 1: {
 				trace("*%ld = ", (p) - r);
-				int val1 = getval(r, *(p++), &pmode);
+				long val1 = getval(r, *(p++), &pmode);
 				trace("+");
-				int val2 = getval(r, *(p++), &pmode);
+				long val2 = getval(r, *(p++), &pmode);
 				trace("\n");
 				r[*(p++)] = val1 + val2;
 				break;
 			}
 			case 2: {
-				int val1 = getval(r, *(p++), &pmode);
-				int val2 = getval(r, *(p++), &pmode);
+				long val1 = getval(r, *(p++), &pmode);
+				long val2 = getval(r, *(p++), &pmode);
 				r[*(p++)] = val1 * val2;
 				break;
 			}
 			case 3: {
 				char *in = NULL;
 				size_t n = 0;
-				trace("*%d = input()(", *p);
-				int status = getline(&in, &n, input);
+				trace("*%ld = input()(", *p);
+				long status = getline(&in, &n, input);
 				if (status == -1) {
 					eprintf("expecting input!\n");
 					exit(1);
 				}
-				trace("=%d)\n", atoi(in));
+				trace("=%ld)\n", atol(in));
 
 				r[*(p++)] = atoi(in);
 				if (in) free(in);
@@ -73,15 +73,15 @@ void computer(int *r, FILE *input) {
 			}
 			case 4: {
 				trace("output(");
-				printf("%d\n", getval(r, *(p++), &pmode));
+				printf("%ld\n", getval(r, *(p++), &pmode));
 				trace(")\n");
 				break;
 			}
 			case 5: {
 				trace("if ");
-				int cond = getval(r, *(p++), &pmode);
+				long cond = getval(r, *(p++), &pmode);
 				trace("? jump");
-				int *loc = r + getval(r, *(p++), &pmode);
+				long *loc = r + getval(r, *(p++), &pmode);
 				trace("\n");
 				trace("loc: %ld\n", loc - r);
 				if (cond!=0) {
@@ -90,22 +90,22 @@ void computer(int *r, FILE *input) {
 				break;
 			}
 			case 6: {
-				int cond = getval(r, *(p++), &pmode);
-				int *loc = r + getval(r, *(p++), &pmode);
+				long cond = getval(r, *(p++), &pmode);
+				long *loc = r + getval(r, *(p++), &pmode);
 				if (cond==0) {
 					p = loc;
 				}
 				break;
 			}
 			case 7: {
-				int val1 = getval(r, *(p++), &pmode);
-				int val2 = getval(r, *(p++), &pmode);
+				long val1 = getval(r, *(p++), &pmode);
+				long val2 = getval(r, *(p++), &pmode);
 				r[*(p++)] = val1 < val2;
 				break;
 			}
 			case 8: {
-				int val1 = getval(r, *(p++), &pmode);
-				int val2 = getval(r, *(p++), &pmode);
+				long val1 = getval(r, *(p++), &pmode);
+				long val2 = getval(r, *(p++), &pmode);
 				r[*(p++)] = val1 == val2;
 				break;
 			}
@@ -114,18 +114,18 @@ void computer(int *r, FILE *input) {
 				break;
 			}
 			default: {
-				eprintf("Unknown op!: %d\n", op);
+				eprintf("Unknown op!: %ld\n", op);
 				exit(0);
 			}
 		}
 		trace("pc: %ld\n", (p) - r);
-		trace("op: %d\n", *p);
+		trace("op: %ld\n", *p);
 	}
 }
 
 
 int main(int argc, char *argv[]) {
-	int reel[10000] = {
+	long reel[10000] = {
 		#ifdef INTPROG
 			INTPROG
 		#else
